test: split Qoder dashboard action tests
